+++ tests/unit_tests/extrinsics/test_serving.py
@@ -21,7 +21,7 @@ import pytest
 from bittensor_wallet import Wallet
 
 from bittensor.core.axon import Axon
-from bittensor.extrinsics.serving import (
+from bittensor.api.extrinsics.serving import (
     serve_extrinsic,
     publish_metadata,
     serve_axon_extrinsic,
@@ -119,7 +119,7 @@ def test_serve_extrinsic_happy_path(
 ):
     # Arrange
     mock_subtensor._do_serve_axon.return_value = (True, "")
-    with patch("bittensor.extrinsics.serving.Confirm.ask", return_value=True):
+    with patch("bittensor.api.extrinsics.serving.Confirm.ask", return_value=True):
         # Act
         result = serve_extrinsic(
             mock_subtensor,
@@ -176,7 +176,7 @@ def test_serve_extrinsic_edge_cases(
 ):
     # Arrange
     mock_subtensor._do_serve_axon.return_value = (True, "")
-    with patch("bittensor.extrinsics.serving.Confirm.ask", return_value=True):
+    with patch("bittensor.api.extrinsics.serving.Confirm.ask", return_value=True):
         # Act
         result = serve_extrinsic(
             mock_subtensor,
@@ -233,7 +233,7 @@ def test_serve_extrinsic_error_cases(
 ):
     # Arrange
     mock_subtensor._do_serve_axon.return_value = (False, "Error serving axon")
-    with patch("bittensor.extrinsics.serving.Confirm.ask", return_value=True):
+    with patch("bittensor.api.extrinsics.serving.Confirm.ask", return_value=True):
         # Act
         result = serve_extrinsic(
             mock_subtensor,
